+++ ft-plugin/src/main/java/com/ft/plugin/FTPlugin.java
@@ -8,6 +8,7 @@ import com.android.build.api.variant.Variant;
 import com.ft.plugin.garble.FTExtension;
 import com.ft.plugin.garble.FTMapUploader;
 import com.ft.plugin.garble.Logger;
+import com.ft.plugin.garble.PluginConfigManager;
 import com.ft.plugin.garble.asm.FTTransform;
 
 import org.gradle.api.Action;
@@ -67,8 +68,11 @@ public class FTPlugin implements Plugin<Project> {
         project.afterEvaluate(p -> {
             //传参数对象
             FTExtension extension = (FTExtension) p.getExtensions().getByName("FTExt");
+            PluginConfigManager.get().setExtension(extension);
 
             Logger.setDebug(extension.showLog);
+            Logger.debug("Plugin Version:" + BuildConfig.PLUGIN_VERSION +
+                    ",ASM Version:" + extension.asmVersion);
 
             FTMapUploader f = new FTMapUploader(p, extension);
             f.configMapUpload();