+++ ft-sdk/src/main/java/com/ft/sdk/FTExceptionHandler.java
@@ -1,7 +1,5 @@
 package com.ft.sdk;
 
-import android.util.Log;
-
 import androidx.annotation.NonNull;
 
 import com.ft.sdk.garble.bean.AppState;
@@ -17,7 +15,6 @@ import java.io.File;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
-import java.io.Writer;
 import java.nio.charset.Charset;
 import java.util.HashMap;
 
@@ -59,6 +56,11 @@ public class FTExceptionHandler implements Thread.UncaughtExceptionHandler {
      * 可以接受到 native crash 回调的版本
      */
     public static final String NATIVE_CALLBACK_VERSION = "1.1.0-alpha01";
+
+    /**
+     * 可以接受 native logcat 行数限制的版本
+     */
+    public static final String NATIVE_LOGCAT_SETTING_VERSION = "1.1.1-alpha01";
 
     private static FTExceptionHandler instance;
     private final Thread.UncaughtExceptionHandler mDefaultExceptionHandler;
@@ -121,7 +123,7 @@ public class FTExceptionHandler implements Thread.UncaughtExceptionHandler {
      */
     @Override
     public void uncaughtException(@NonNull Thread t, @NonNull Throwable e) {
-        Writer writer = new StringWriter();
+        StringWriter writer = new StringWriter();
         PrintWriter printWriter = new PrintWriter(writer);
         e.printStackTrace(printWriter);
         Throwable cause = e.getCause();
@@ -130,8 +132,16 @@ public class FTExceptionHandler implements Thread.UncaughtExceptionHandler {
             cause = cause.getCause();
         }
         printWriter.close();
-        String result = writer.toString();
-        uploadCrashLog(result, e.getMessage(), FTActivityManager.get().getAppState(), new RunnerCompleteCallBack() {
+        writer.append("\n").append(Utils.getAllThreadStack());
+        ExtraLogCatSetting logCatWithError = config.getExtraLogCatWithJavaCrash();
+        if (logCatWithError != null) {
+            writer.append("\n")
+                    .append(Utils.getLogcat(logCatWithError.getLogcatMainLines(),
+                            logCatWithError.getLogcatSystemLines(),
+                            logCatWithError.getLogcatEventsLines()));
+        }
+
+        uploadCrashLog(writer.toString(), e.getMessage(), FTActivityManager.get().getAppState(), new RunnerCompleteCallBack() {
             @Override
             public void onComplete() {
                 //测试用例直接
@@ -177,7 +187,7 @@ public class FTExceptionHandler implements Thread.UncaughtExceptionHandler {
                                 uploadNativeCrash(item, AppState.getValueFrom(value), true, callBack);
                                 Utils.deleteFile(item.getAbsolutePath());
                             } catch (IOException e) {
-                                LogUtils.e(TAG, Log.getStackTraceString(e));
+                                LogUtils.e(TAG, LogUtils.getStackTraceString(e));
                             }
                         }
                     }
@@ -204,7 +214,7 @@ public class FTExceptionHandler implements Thread.UncaughtExceptionHandler {
                 try {
                     uploadNativeCrash(item, state, isPreCrash, callBack);
                 } catch (Exception e) {
-                    LogUtils.e(TAG, Log.getStackTraceString(e));
+                    LogUtils.e(TAG, LogUtils.getStackTraceString(e));
                     if (callBack != null) {
                         callBack.onComplete();
                     }