+++ ft-sdk/src/main/java/com/ft/sdk/garble/http/NativeNetEngine.java
@@ -2,8 +2,6 @@ package com.ft.sdk.garble.http;
 
 import static com.ft.sdk.garble.http.NetCodeStatus.UNKNOWN_EXCEPTION_CODE;
 
-import android.util.Log;
-
 import com.ft.sdk.garble.FTHttpConfigManager;
 import com.ft.sdk.garble.utils.Constants;
 import com.ft.sdk.garble.utils.LogUtils;
@@ -94,7 +92,7 @@ public class NativeNetEngine implements INetEngine {
                 connSuccess = true;
             }
         } catch (Exception e) {
-            LogUtils.e(TAG, Log.getStackTraceString(e));
+            LogUtils.e(TAG, LogUtils.getStackTraceString(e));
 
         }
         return connSuccess;
@@ -108,7 +106,7 @@ public class NativeNetEngine implements INetEngine {
             //设置连接方式
             mConnection.setRequestMethod(mHttpBuilder.getMethod().method);
         } catch (ProtocolException e) {
-            LogUtils.e(TAG, Log.getStackTraceString(e));
+            LogUtils.e(TAG, LogUtils.getStackTraceString(e));
         }
         //设置连接和读取超时时间
         mConnection.setConnectTimeout(FTHttpConfigManager.get().getSendOutTime());
@@ -217,28 +215,28 @@ public class NativeNetEngine implements INetEngine {
                 outputStream.close();
             }
         } catch (IOException e) {
-            LogUtils.e(TAG, Log.getStackTraceString(e));
+            LogUtils.e(TAG, LogUtils.getStackTraceString(e));
         }
         try {
             if (reader != null) {
                 reader.close();
             }
         } catch (IOException e) {
-            LogUtils.e(TAG, Log.getStackTraceString(e));
+            LogUtils.e(TAG, LogUtils.getStackTraceString(e));
         }
         try {
             if (inputStreamReader != null) {
                 inputStreamReader.close();
             }
         } catch (IOException e) {
-            LogUtils.e(TAG, Log.getStackTraceString(e));
+            LogUtils.e(TAG, LogUtils.getStackTraceString(e));
         }
         try {
             if (inputStream != null) {
                 inputStream.close();
             }
         } catch (IOException e) {
-            LogUtils.e(TAG, Log.getStackTraceString(e));
+            LogUtils.e(TAG, LogUtils.getStackTraceString(e));
 
         }
     }