+++ ft-sdk/src/main/java/com/ft/sdk/internal/anr/ANRDetectRunnable.java
@@ -4,12 +4,14 @@ package com.ft.sdk.internal.anr;
 import android.os.Handler;
 import android.os.Looper;
 
+import com.ft.sdk.ExtraLogCatSetting;
 import com.ft.sdk.FTRUMGlobalManager;
 import com.ft.sdk.garble.bean.AppState;
 import com.ft.sdk.garble.bean.ErrorType;
 import com.ft.sdk.garble.utils.Constants;
 import com.ft.sdk.garble.utils.LogUtils;
 import com.ft.sdk.garble.utils.StringUtils;
+import com.ft.sdk.garble.utils.Utils;
 
 /**
  * ANR 错误监测，循环监测两个 Runner 执行相差时间是否超过 {@link #ANR_DETECT_DURATION_MS},超过则追加一条 ANR 错误信息
@@ -23,12 +25,15 @@ public final class ANRDetectRunnable implements Runnable {
      */
     public static final int ANR_DETECT_DURATION_MS = 5000;
 
+    private final ExtraLogCatSetting extraLogCatSetting;
+
     /**
      * 主线程消息 Handler
      */
     private final Handler handler = new Handler(Looper.getMainLooper());
 
-    public ANRDetectRunnable() {
+    public ANRDetectRunnable(ExtraLogCatSetting extraLogCatSetting) {
+        this.extraLogCatSetting = extraLogCatSetting;
     }
 
     private final CallbackRunnable runnable = new CallbackRunnable();
@@ -48,15 +53,20 @@ public final class ANRDetectRunnable implements Runnable {
                     runnable.wait(ANR_DETECT_DURATION_MS);
 
                     if (!runnable.isCalled()) {
+                        String stackTrace =
+                                StringUtils.getStringFromStackTraceElement(handler.getLooper().getThread().getStackTrace())
+                                        + "\n" + Utils.getAllThreadStack();
+                        if (extraLogCatSetting != null) {
+                            stackTrace += Utils.getLogcat(extraLogCatSetting.getLogcatMainLines(),
+                                    extraLogCatSetting.getLogcatSystemLines(),
+                                    extraLogCatSetting.getLogcatEventsLines());
+                        }
+
                         //如果超时间没有调用则，添加一条 Error 数据
-                        FTRUMGlobalManager.get().addError(
-                                StringUtils.getStringFromStackTraceElement(handler.getLooper().getThread().getStackTrace()),
-                                "android_anr", ErrorType.ANR_ERROR, AppState.RUN);
+                        FTRUMGlobalManager.get().addError(stackTrace, "android_anr", ErrorType.ANR_ERROR, AppState.RUN);
                         runnable.wait();
                     }
                 }
-
-
             } catch (InterruptedException e) {
                 LogUtils.e(TAG, "ANR Thread interrupt");
                 break;
@@ -79,6 +89,7 @@ public final class ANRDetectRunnable implements Runnable {
     public static class CallbackRunnable implements Runnable {
         /**
          * 是否被调用
+         *
          * @return
          */
         public boolean isCalled() {